server/requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
certifi==14.5.14
defusedxml==0.4.1
Django==1.9.5
django-allauth==0.25.2
django-autoslug==1.9.3
django-bootstrap3==7.0.1
django-ckeditor==4.4.7
django-colorful==1.2
django-comments-xtd==1.5.3
django-contrib-comments==1.7.0
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-fab-deploy==0.7.5
django-favit==0.2.4
django-filter==0.9.2
django-haystack==2.4.1
django-jsonify==0.3.0
django-paintstore==0.2
django-pgjson==0.2.3
django-push-notifications==1.3.1
django-rest-auth==0.7.0
django-taggit==0.18.1
django-termsandconditions==0.1.6
django-uuidfield==0.5.0
djangorestframework==3.3.3
djmail==0.10.0
docutils==0.12
drf-haystack==1.2
ecdsa==0.13
elasticsearch==1.4.0
Fabric==1.11.1
geopy==1.9.1
ipdb==0.8.1
ipython==2.4.1
Jinja2==2.7.3
jsonschema==2.4.0
kombu==3.0.26
MarkupSafe==0.23
mistune==0.5
nose==1.3.4
numpydoc==0.5
oauthlib==1.0.3
paramiko==1.16.0
Pillow==2.7.0
psycopg2==2.6
ptyprocess==0.4
pycrypto==2.6.1
pyelasticsearch==1.2
Pygments==2.0.2
python3-openid==3.0.10
pytz==2014.10
pyzmq==14.5.0
requests==2.9.1
requests-oauthlib==0.6.1
simplejson==3.6.5
six==1.10.0
sorl-thumbnail==12.4a1
Sphinx==1.2.3
sqlparse==0.1.19
terminado==0.5
tornado==4.1
urllib3==1.10.2
dj-database-url
gunicorn
#django-celery